move Avatar components to their own folder
